Refactor Brave Ads timestamps from uint64_t to double and base::Time

+++ browser/ui/webui/brave_rewards_page_ui.cc
@@ -128,7 +128,7 @@ class RewardsDOMHandler : public WebUIMessageHandler,
   void GetExcludedSites(const base::ListValue* args);
 
   void OnGetStatement(const bool success,
-                      const int64_t next_payment_date,
+                      const double next_payment_date,
                       const int ads_received_this_month,
                       const double earnings_this_month,
                       const double earnings_last_month);
@@ -1441,7 +1441,7 @@ void RewardsDOMHandler::GetStatement(
 }
 
 void RewardsDOMHandler::OnGetStatement(const bool success,
-                                       const int64_t next_payment_date,
+                                       const double next_payment_date,
                                        const int ads_received_this_month,
                                        const double earnings_this_month,
                                        const double earnings_last_month) {
